pyrad/client: use select.poll() instead of select.select()
select.select() has a limitation of not being able to use file descriptors
larger than FD_SETSIZE. This causes issues when trying to use pyrad.Client
in a long-running software that is keeping a large number of connections
open.

This change converts Client to use poll() instead of select() to wait for
an incoming packet.